requirements.txt
argparse==1.4.0
pyelftools==0.27
